Moved IKPoints.cs and AnimationController.cs to _Project/Scripts/Core/Character/Animation folder
