+++ Assets/_Project/Scripts/Core/Character/Animation.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f899bc1981bc4813b4a95c324ab60140
+timeCreated: 1741101828
