+++ Assets/_Project/Scripts/Core/Character/Animation/IKPoints.cs
@@ -1,9 +1,8 @@
-using System;
 using _Project.Scripts.Core.Character.Weapon_Controller;
 using UnityEngine;
 using UnityEngine.Animations.Rigging;
 
-namespace _Project.Scripts.Core.Character.IK_Points
+namespace _Project.Scripts.Core.Character.Animation
 {
     public class IKPoints : MonoBehaviour
     {
