--- Assets/_Project/Scripts/Core/Character/IK Points.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0324248a6cb3243d1bd9900cbe57ed84
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
